Fixed build bug & started thickening-view
Bug was related to the texture not being loaded correctly, as it was changed to a SerializedProperty for debugging pusposes in the editor and that caused it to not show up in the build.
Also worked on shifting vertices depending on the view to make grass look dense from all sides. Not working correctly yet.